backend/requirements.txt
fastapi
uvicorn
pillow
svgwrite